refactor: move services to be it's own module in package
