+++ src/brisk/services/bundle.py
@@ -2,10 +2,10 @@
 
 import dataclasses
 
-from brisk.evaluation.services.logging import LoggingService
-from brisk.evaluation.services.metadata import MetadataService
-from brisk.evaluation.services.io import IOService
-from brisk.evaluation.services.utility import UtilityService
+from brisk.services.logging import LoggingService
+from brisk.services.metadata import MetadataService
+from brisk.services.io import IOService
+from brisk.services.utility import UtilityService
 
 @dataclasses.dataclass
 class ServiceBundle:
